fixed gallery pagers
